requirements.txt
google-generativeai>=0.3.0
numpy>=1.24.0
scipy>=1.10.0
tiktoken>=0.5.0
pytest>=7.4.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pydantic>=2.0.0

